Remove old spark-hdfs from code base
