+++ sparkcore/src/test/scala/quasar/physical/sparkcore/fs/hdfs/HdfsPackageSpec.scala
@@ -39,6 +39,6 @@ class HdfsPackageSpec extends quasar.Qspec {
   }
 
   private def hdfsUri(url: String): Task[URI] =
-    generateHdfsFS(SparkFSConf(new SparkConf(), url, rootDir)).map(_.getUri())
+    SparkHdfsBackendModule.generateHdfsFS(HdfsConfig(new SparkConf(), url, rootDir)).map(_.getUri())
 
 }
